move all timeline-returning functions to Timelines.kt
- renaming all functions to get***Timeline to be consistent
- moving around related functions in Rx and tests

Closes #81
