+++ bigbone/src/main/kotlin/social/bigbone/api/method/Timelines.kt
@@ -11,14 +11,94 @@ import social.bigbone.api.exception.BigboneRequestException
  * see more https://github.com/tootsuite/documentation/blob/master/Using-the-API/API.md#timelines
  */
 class Timelines(private val client: MastodonClient) {
-    //  GET /api/v1/timelines/home
+    /**
+     * Public timelines can consist of only local statuses, only remote (=federated) statuses, or a combination of both.
+     */
+    enum class StatusOrigin {
+        LOCAL,
+        REMOTE,
+        LOCAL_AND_REMOTE
+    }
+
+    /**
+     * Gets the home timeline of statuses.
+     * @param range restrict result to a specific range
+     * @see <a href="https://docs.joinmastodon.org/methods/timelines/#home">Mastodon API documentation: methods/timelines/#home</a>
+     */
     @JvmOverloads
     @Throws(BigboneRequestException::class)
-    fun getHome(range: Range = Range()): MastodonRequest<Pageable<Status>> {
+    fun getHomeTimeline(range: Range = Range()): MastodonRequest<Pageable<Status>> {
         return client.getPageableMastodonRequest(
             endpoint = "api/v1/timelines/home",
             method = MastodonClient.Method.GET,
             parameters = range.toParameter()
+        )
+    }
+
+    /**
+     * Gets a list timeline of statuses for the given list.
+     * @param listID ID of the list for which a timeline should be returned
+     * @param range restrict result to a specific range
+     * @see <a href="https://docs.joinmastodon.org/methods/timelines/#list">Mastodon API documentation: methods/timelines/#list</a>
+     */
+    @Throws(BigboneRequestException::class)
+    fun getListTimeline(listID: String, range: Range = Range()): MastodonRequest<Pageable<Status>> {
+        return client.getPageableMastodonRequest(
+            endpoint = "api/v1/timelines/list/$listID",
+            method = MastodonClient.Method.GET,
+            parameters = range.toParameter()
+        )
+    }
+
+    /**
+     * Get the public timeline of the configured instance. Defaults to a combination of local and remote statuses,
+     * but can be restricted to either.
+     * @param range restrict result to a specific range
+     * @param statusOrigin optionally restrict result to either local or remote (=federated) statuses; defaults to all
+     * @see <a href="https://docs.joinmastodon.org/methods/timelines/#public">Mastodon API documentation: methods/timelines/#public</a>
+     */
+    @JvmOverloads
+    fun getPublicTimeline(
+        range: Range = Range(),
+        statusOrigin: StatusOrigin = StatusOrigin.LOCAL_AND_REMOTE
+    ): MastodonRequest<Pageable<Status>> {
+        return client.getPageableMastodonRequest(
+            endpoint = "api/v1/timelines/public",
+            method = MastodonClient.Method.GET,
+            parameters = range.toParameter().apply {
+                when (statusOrigin) {
+                    StatusOrigin.LOCAL -> append("local", true)
+                    StatusOrigin.REMOTE -> append("remote", true)
+                    else -> { /* append neither, combined results will be shown */ }
+                }
+            }
+        )
+    }
+
+    /**
+     * Get the public timeline for a specific hashtag of the configured instance.
+     * Defaults to a combination of local and remote (=federated) statuses, but can be restricted to either.
+     * @param tag the hashtag for which a timeline should be returned
+     * @param range restrict result to a specific range
+     * @param statusOrigin optionally restrict result to either local or remote statuses; defaults to all
+     * @see <a href="https://docs.joinmastodon.org/methods/timelines/#tag">Mastodon API documentation: methods/timelines/#tag</a>
+     */
+    @JvmOverloads
+    fun getTagTimeline(
+        tag: String,
+        range: Range = Range(),
+        statusOrigin: StatusOrigin = StatusOrigin.LOCAL_AND_REMOTE
+    ): MastodonRequest<Pageable<Status>> {
+        return client.getPageableMastodonRequest(
+            endpoint = "api/v1/timelines/tag/$tag",
+            method = MastodonClient.Method.GET,
+            parameters = range.toParameter().apply {
+                when (statusOrigin) {
+                    StatusOrigin.LOCAL -> append("local", true)
+                    StatusOrigin.REMOTE -> append("remote", true)
+                    else -> { /* append neither, combined results will be shown */ }
+                }
+            }
         )
     }
 }
