+++ sample-java/src/main/java/social/bigbone/sample/GetPublicTimelines.java
@@ -6,16 +6,16 @@ import social.bigbone.api.Range;
 import social.bigbone.api.entity.Account;
 import social.bigbone.api.entity.Status;
 import social.bigbone.api.exception.BigboneRequestException;
-import social.bigbone.api.method.Public;
+import social.bigbone.api.method.Timelines;
 
 @SuppressWarnings({"PMD.AvoidPrintStackTrace", "PMD.SystemPrintln"})
 public class GetPublicTimelines {
     public static void main(final String[] args) {
         final MastodonClient client = new MastodonClient.Builder("mstdn.jp").build();
-        final Public publicMethod = new Public(client);
+        final Timelines timelines = new Timelines(client);
 
         try {
-            final Pageable<Status> statuses = publicMethod.getPublic(new Range(), Public.StatusOrigin.LOCAL)
+            final Pageable<Status> statuses = timelines.getPublicTimeline(new Range(), Timelines.StatusOrigin.LOCAL)
                     .doOnJson(System.out::println)
                     .execute();
             statuses.getPart().forEach(status -> {
