+++ sample-java/src/main/java/social/bigbone/sample/GetRawJson.java
@@ -3,7 +3,7 @@ package social.bigbone.sample;
 import social.bigbone.MastodonClient;
 import social.bigbone.api.Range;
 import social.bigbone.api.exception.BigboneRequestException;
-import social.bigbone.api.method.Public;
+import social.bigbone.api.method.Timelines;
 
 import java.io.IOException;
 
@@ -14,8 +14,8 @@ public class GetRawJson {
             final String instanceName = args[0];
             final String credentialFilePath = args[1];
             final MastodonClient client = Authenticator.appRegistrationIfNeeded(instanceName, credentialFilePath, false);
-            final Public publicMethod = new Public(client);
-            publicMethod.getPublic(new Range(), Public.StatusOrigin.LOCAL).doOnJson(System.out::println).execute();
+            final Timelines timelines = new Timelines(client);
+            timelines.getPublicTimeline(new Range(), Timelines.StatusOrigin.LOCAL).doOnJson(System.out::println).execute();
         } catch (IOException | BigboneRequestException e) {
             e.printStackTrace();
         }
