+++ sample-java/src/main/java/social/bigbone/sample/GetTagTimelines.java
@@ -5,17 +5,17 @@ import social.bigbone.api.Pageable;
 import social.bigbone.api.Range;
 import social.bigbone.api.entity.Account;
 import social.bigbone.api.entity.Status;
-import social.bigbone.api.method.Public;
 import social.bigbone.api.exception.BigboneRequestException;
+import social.bigbone.api.method.Timelines;
 
 @SuppressWarnings({"PMD.SystemPrintln", "PMD.AvoidPrintStackTrace"})
 public class GetTagTimelines {
     public static void main(final String[] args) {
         final MastodonClient client = new MastodonClient.Builder("mstdn.jp").build();
-        final Public publicMethod = new Public(client);
+        final Timelines timelines = new Timelines(client);
 
         try {
-            final Pageable<Status> statuses = publicMethod.getTag("mastodon", new Range(), Public.StatusOrigin.LOCAL_AND_REMOTE).execute();
+            final Pageable<Status> statuses = timelines.getTagTimeline("mastodon", new Range(), Timelines.StatusOrigin.LOCAL_AND_REMOTE).execute();
             statuses.getPart().forEach(status -> {
                 System.out.println("=============");
                 final Account account = status.getAccount();
