+++ sample-kotlin/src/main/kotlin/social/bigbone/sample/GetRawJson.kt
@@ -1,7 +1,7 @@
 package social.bigbone.sample
 
 import kotlinx.coroutines.runBlocking
-import social.bigbone.api.method.Public
+import social.bigbone.api.method.Timelines
 
 object GetRawJson {
     @JvmStatic
@@ -10,8 +10,8 @@ object GetRawJson {
         val credentialFilePath = args[1]
         val client = Authenticator.appRegistrationIfNeeded(instanceName, credentialFilePath)
         runBlocking {
-            val public = Public(client)
-            public.getPublic(statusOrigin = Public.StatusOrigin.LOCAL).doOnJson {
+            val timelines = Timelines(client)
+            timelines.getPublicTimeline(statusOrigin = Timelines.StatusOrigin.LOCAL).doOnJson {
                 println(it)
             }.execute()
         }
