+++ sample-kotlin/src/main/kotlin/social/bigbone/sample/Kotlindon.kt
@@ -16,17 +16,17 @@ object Kotlindon {
         val instanceName = args[0]
         val credentialFilePath = args[1]
         val client = Authenticator.appRegistrationIfNeeded(instanceName, credentialFilePath)
-        listenHome(client)
+        listenHomeTimeline(client)
     }
 
-    fun listenHome(client: MastodonClient) {
+    fun listenHomeTimeline(client: MastodonClient) {
         runBlocking {
             val timelines = Timelines(client)
             var pageable: Pageable<Status>? = null
             while (true) {
                 val result = pageable?.let {
-                    timelines.getHome(it.prevRange(limit = 5)).execute()
-                } ?: timelines.getHome(Range(limit = 5)).execute()
+                    timelines.getHomeTimeline(it.prevRange(limit = 5)).execute()
+                } ?: timelines.getHomeTimeline(Range(limit = 5)).execute()
 
                 result.part.sortedBy { it.createdAt }.forEach {
                     println(it.account?.displayName)
